character: accumulate in i128 and report overflow, instead of wrapping silently
symfn's Murnaghan-Nakayama characters accumulated in i64. |chi^lambda(mu)| is
bounded by d_lambda, and sum of d_lambda^2 = n! puts max d_lambda near
sqrt(n!), which passes i64 at n ~ 35. Past that, release builds wrapped and
returned a wrong number with no indication:

  n     symfn i64                hook-length i128
  34    427940616636149760       427940616636149760      ok
  36    -7635121151160614912     29258366996258488320    WRONG
  40    6142810650918140564      23820889409809949276820 WRONG

At n=36 it reported a *negative dimension*, which is impossible. Nothing
caught it: the existing tests stop at n=8, and debug builds would trap but no
test reaches the size. The corrupted values fed straight into the s <-> p
transitions in convert.rs and into character_value in the Python bridge.

Changes:

  - character() returns i128, accumulating with checked_add. Beyond i128
    (n ~ 58) it panics with a message naming the shape; try_character() returns
    None instead of panicking. Panicking is deliberate -- the thing it
    replaces is a silently wrong answer flowing into transition matrices where
    right and wrong look identical.
  - Ring::from_i128, mirroring the existing from_u128 seam, so convert.rs
    injects characters without truncating. Exact for i128 and for rug::Integer
    / rug::Rational under the gmp feature.
  - character_cached now caches i128 and deliberately does NOT cache overflow:
    None is the absence of a representable value, not a value, and caching it
    would leave the table unable to tell "uncomputed" from "unrepresentable".

Tests: the identity class is checked against the hook-length formula -- an
independent computation, not the same recursion -- across n = 20..48, which
straddles the n=36 boundary and would have caught the original bug. A second
test pins that overflow past i128 is reported rather than wrapped.

The scale of the fix is bounded: i128 covers to n ~ 58. Making characters
generic over Ring, so the gmp feature lifts the ceiling entirely, is the
principled follow-up and is noted for later -- it would also close the related
gap that python.rs hardcodes i64 coefficients, so gmp cannot reach Sage.

Found by comparing against an independent M-N implementation the author had
written elsewhere, which had chosen i128 with checked arithmetic for exactly
this reason.

+++ src/python.rs
@@ -152,7 +152,7 @@ fn kostka_number(lambda: Vec<u32>, mu: Vec<u32>) -> u128 {
 
 /// Symmetric-group character χ^λ(μ).
 #[pyfunction]
-fn character_value(lambda: Vec<u32>, mu: Vec<u32>) -> i64 {
+fn character_value(lambda: Vec<u32>, mu: Vec<u32>) -> i128 {
     crate::character::character(&part(&lambda), &part(&mu))
 }
